+++ test/integration/slipway/test_server.clj
@@ -4,10 +4,10 @@
             [slipway.context :as context]
             [slipway.example.app :as app]
             [slipway.security :as security]
-            [slipway.security.openid :as openid]
-            [slipway.security.openid.jwks :as openid.jwks]
-            [slipway.security.openid.jwt :as openid.jwt]
-            [slipway.security.openid.jwt.at.verification :as openid.jwt.at.verification]
+            [slipway.security.oidc :as oidc]
+            [slipway.security.oidc.jwks :as oidc.jwks]
+            [slipway.security.oidc.jwt :as oidc.jwt]
+            [slipway.security.oidc.jwt.at.verification :as oidc.jwt.at.verification]
             [slipway.sente]
             [slipway.server :as server]
             [slipway.session :as session]))
@@ -31,29 +31,28 @@
 (defn start-with-openid-auth-code!
   []
   (start! #::server{:connector     {::http/port 3000}
-                    :handler       {::context/ring-handler         (app/handler)
-                                    ::security/handler             :openid
-                                    ::openid/issuer                "http://localhost:8080/realms/master"
-                                    ::openid/client-id             "slipway"
-                                    ::openid/client-secret         "81a0d6ea-1468-4b20-b115-fa68a8df9cf8"
-                                    ;;::openid/scopes                ["profile" "email"] ;; these are the no-config defaults, "openid" is always present regardless.
-                                    ::openid.jwt/user-id-path      ["name"]
-                                    ::openid.jwt/user-roles-path   ["realm_access" "roles"]
+                    :handler       {::context/ring-handler       (app/handler)
+                                    ::security/handler           :oidc
+                                    ::oidc/issuer                "http://localhost:8080/realms/master"
+                                    ::oidc/client-id             "slipway"
+                                    ::oidc/client-secret         "81a0d6ea-1468-4b20-b115-fa68a8df9cf8"
+                                    ::oidc/constraint-mappings   app/constraints
+                                    ::oidc.jwt/user-id-path      ["name"]
 
                                     ;; The following three parameters are derived automatically from the issuer
                                     ;; E.g. http://localhost:8080/realms/master/.well-known/openid-configuration
-                                    ;; Jetty looks them all up when only ::openid/issuer is provided
+                                    ;; Jetty looks them all up when only ::oidc/issuer is provided
                                     ;; Optional specific endpoint testing
-                                    ;::openid/authorization-endpoint "http://localhost:8080/realms/master/protocol/openid-connect/auth"
-                                    ;::openid/token-endpoint         "http://localhost:8080/realms/master/protocol/openid-connect/token"
-                                    ;::openid/end-session-endpoint   "http://localhost:8080/realms/master/protocol/openid-connect/logout"
+                                    ;::oidc/authorization-endpoint "http://localhost:8080/realms/master/protocol/openid-connect/auth"
+                                    ;::oidc/token-endpoint         "http://localhost:8080/realms/master/protocol/openid-connect/token"
+                                    ;::oidc/end-session-endpoint   "http://localhost:8080/realms/master/protocol/openid-connect/logout"
 
                                     ;; Optional redirect testing
-                                    ::openid/oidc-redirect-success "/oauth2/openid/callback" ;; defaults to /j_security_check but is configurable
-                                    ::openid/oidc-redirect-error   "/login-error" ;; `http://localhost:3000/login-error?error_description_jetty=ID+Token+has+expired` when token expired mid-auth-flow
-                                    ::openid/oidc-redirect-logout  "/logout-success" ;; redirected to post-logout
+                                    ::oidc.jwt/user-roles-path   ["realm_access" "roles"] ;; defaults to /j_security_check but is configurable
+                                    ::oidc/oidc-redirect-success "/oauth2/openid/callback" ;; `http://localhost:3000/login-error?error_description_jetty=ID+Token+has+expired` when token expired mid-auth-flow
+                                    ::oidc/oidc-redirect-error   "/login-error" ;; redirected to post-logout
 
-                                    ::openid/constraint-mappings   app/constraints}
+                                    ::oidc/oidc-redirect-logout  "/logout-success"}
                     :error-handler app/server-error-handler}))
 
 ;; ./scripts/keycloak.sh
@@ -62,14 +61,14 @@
 (defn start-with-openid-client-creds!
   []
   (start! #::server{:connector     {::http/port 3000}
-                    :handler       {::context/ring-handler                 (app/handler)
-                                    ::security/handler                     :openid
-                                    ::session/enabled?                     false
-                                    ::openid/authorization-flow            :client-credentials
-                                    ::openid.jwks/endpoint                 "http://localhost:8080/realms/master/protocol/openid-connect/certs"
-                                    ::openid.jwt.at.verification/exact-iss "http://localhost:8080/realms/master"
-                                    ::openid.jwt.at.verification/exact-aud "https://slipway.io/api" ;; <-- set in keycloak-realms-with-client.json
-                                    ::openid.jwt/user-id-path              ["preferred_username"]
-                                    ::openid.jwt/user-roles-path           ["realm_access" "roles"]
-                                    ::openid/constraint-mappings           app/constraints}
+                    :handler       {::context/ring-handler               (app/handler)
+                                    ::security/handler                   :oidc
+                                    ::session/enabled?                   false
+                                    ::oidc/authorization-flow            :client-credentials
+                                    ::oidc/constraint-mappings           app/constraints
+                                    ::oidc.jwks/endpoint                 "http://localhost:8080/realms/master/protocol/openid-connect/certs"
+                                    ::oidc.jwt.at.verification/exact-iss "http://localhost:8080/realms/master" ;; <-- set in keycloak-realms-with-client.json
+                                    ::oidc.jwt.at.verification/exact-aud "https://slipway.io/api"
+                                    ::oidc.jwt/user-id-path              ["preferred_username"]
+                                    ::oidc.jwt/user-roles-path           ["realm_access" "roles"]}
                     :error-handler app/server-error-handler}))