+++ test/unit/slipway/request_test.clj
@@ -2,18 +2,18 @@
   (:require [clojure.test :refer [deftest is]]
             [slipway.principal :as-alias principal]
             [slipway.request :as request]
-            [slipway.security.openid :as-alias openid]
+            [slipway.security.oidc :as-alias oidc]
             [slipway.user :as user])
   (:import (java.time Instant)))
 
 (deftest request-user-state
 
-  (let [request-user {::user/identity {::principal/type  ::openid/principal
+  (let [request-user {::user/identity {::principal/type  ::oidc/principal
                                        ::principal/name  "my-user-id"
                                        ::user/roles      ["my-role-1" "my-role-2"]
                                        ::user/expires-at (Instant/ofEpochSecond 1311281970)}}]
 
-    (is (= ::openid/principal (request/user-type request-user)))
+    (is (= ::oidc/principal (request/user-type request-user)))
     (is (= "my-user-id" (request/user-name request-user)))
     (is (= ["my-role-1" "my-role-2"] (request/user-roles request-user)))
     (is (request/user-expired? request-user))))